% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusteredHeatmap.R
\name{clusteredHeatmap}
\alias{clusteredHeatmap}
\title{clusteredHeatmap}
\usage{
clusteredHeatmap(condition1, condition2, top_count = 50)
}
\arguments{
\item{condition1}{Reference condition}

\item{condition2}{Experimental condition}

\item{top_count}{How many genes to plot}
}
\value{
A ComplexHeatmap object
}
\description{
This function returns a clustered heatmap of the top N genes by adjusted p-value
}
