% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compDESeq2.R
\name{compDESeq2}
\alias{compDESeq2}
\title{compDESeq2}
\usage{
compDESeq2(condition1, condition2)
}
\arguments{
\item{condition1}{Reference condition}

\item{condition2}{Experimental condition}
}
\value{
A data.table object
}
\description{
This function returns a merged data.table object from a DESeq2 results function.
}
