% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTx2gene.R
\name{getTx2gene}
\alias{getTx2gene}
\title{Get a tx2gene dataframe for the given species}
\usage{
getTx2gene(species)
}
\arguments{
\item{species}{A valid ENSEMBL species abbreviation e.g. \verb{mmusculus, rnorvegicus, hsapiesn}}
}
\value{
A dataframe of the tx2gene object
}
\description{
This function returns a tx2gene dataframe object for the queried species
}
