% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot3DPCA.R
\name{plot3DPCA}
\alias{plot3DPCA}
\title{Plot PCA Data in 3D}
\usage{
plot3DPCA(object, intgroup = "condition", ntop = 500)
}
\arguments{
\item{object}{a ‘DESeqDataSet’ object with results metadata after running \code{DESeq}.}

\item{intgroup}{interesting groups: a character vector of names in ‘colData(x)’ to use for grouping}

\item{ntop}{number of top genes to use for principal components, selected by highest row variance}
}
\value{
A plotly object of the PCA data
}
\description{
This function takes a ‘DESeqDataSet’ object and returns a 3D PCA plot.
}
