% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVolcano.R
\name{plotVolcano}
\alias{plotVolcano}
\title{Plot Volcano}
\usage{
plotVolcano(condition1, condition2)
}
\arguments{
\item{condition1}{Reference condition}

\item{condition2}{Experimental condition}
}
\value{
A ggplot of the volcano plot
}
\description{
This function returns a volcano plot of the supplied conditions from a DESeq2 object.
}
