% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returnPCA.R
\name{returnPCA}
\alias{returnPCA}
\title{Return PCA Data}
\usage{
returnPCA(object = dds, intgroup = "condition", ntop = 500)
}
\arguments{
\item{object}{a ‘DESeqDataSet’ object with results metadata after running \code{DESeq}.}

\item{intgroup}{interesting groups: a character vector of names in ‘colData(x)’ to use for grouping}

\item{ntop}{number of top genes to use for principal components, selected by highest row variance}
}
\value{
A dataframe of the PCA data
}
\description{
This function takes a ‘DESeqDataSet’ object object, performs a variance stabilizing transform, and returns PCA data for PC1-4 as a dataframe. This code is modified from the DESeq2 function 'plotPCA'.
}
